СИ/requirements.txt
numpy>=2.1
pandas
openpyxl>=3.1
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary>=2.9.9
python-dotenv==1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
